src/requirements.txt
zerosms
twilio
Django==2.2.2
djangorestframework==3.9.4
Pillow==6.0.0
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
